[Opcode/Struct]: SMSG_MOVE_SET_ACTIVE_MOVER + SMSG_MOVE_KNOCK_BACK + SMSG_MOVE_SET_HOVER + SMSG_MOVE_UNSET_HOVER + SMSG_SPLINE_MOVE_GRAVITY_DISABLE + SMSG_SPLINE_MOVE_GRAVITY_ENABLE + SMSG_SPLINE_MOVE_SET_RUN_MODE + SMSG_SPLINE_MOVE_SET_WALK_MODE

+++ src/server/game/Spells/SpellEffects.cpp
@@ -2588,7 +2588,7 @@ void Spell::EffectEnergize(SpellEffIndex effIndex)
         {
             damage = CalculatePct(m_caster->GetMaxHealth(), damage);
 
-            if (m_caster->GetHealth() < damage)
+            if (int32(m_caster->GetHealth()) < damage)
             {
                 damage = m_caster->GetHealth() - 1;
                 m_caster->SetHealth(1);
